feat(be): 댓글 등록/수정/삭제 API 인증/인가 추가 (#157) (#170)
* Test: 댓글등록 테스트 수정

- @WithUserDetails 추가
- JWT AccessToken을 헤더로 등록
 Please enter the commit message for your changes. Lines starting

* Feat: 시큐리티 설정 추가

- 댓글 등록시 인증 필요

* Feat: 댓글 등록 인증/인가 기능 추가

- @RequestHeader로 accessToken 받아옴
- accessToken에서 userId를 파싱
- service에서 User 정보를 등록

* Feat(be): 시큐리티 설정 추가

- 댓글 수정/삭제 API 인증/인가 필요하도록

* Feat: 댓글 수정/삭제 권한 에러메시지 추가

- ErrorCode COMMENT_002, COMMENT_003 추가

* Feat: 댓글 수정/삭제 권한 확인 로직 추가

- AccessToken을 받아온 후, 댓글과 토큰의 userId를 비교
- 만약 두 아이디가 다른 경우 예외처리

* Test: 댓글 등록/수정/삭제 테스트 코드 인증/인가 추가

인증/인가 로직이 추가됨에 따라 테스트 코드 변경

- 각 테스트 user1의 accessToken에 추가
- Header에 accessToken 추가설정

* Fix: 댓글관련 시큐리티 설정 삭제

- SecurityConfig에 댓글관련 시큐리티 설정 삭제

* Fix: 댓글 등록/삭제/수정 Controller 메서드 수정

- @PreAuthorization 추가
- 기존 -> Header로부터 Authorization 정보를 받아오는 방식
- 변경 -> CustomUserDetails로 사용자 정보를 받아오는 방식

* Test: CommentControllerTest 수정
- AccessToken 삭제
- 유틸 메서드 삭제
- Header 부분 삭제

* Feat: 시큐리티 설정추가

- 댓글 조회의 경우 인증 없이도 사용가능
- 그 이외의 댓글기능은 인증 후, 사용가능

* Fix: 유저 정보 받는 방식 변경

기존 -> CustomUserDetails를 사용
변경 -> Rq를 사용

+++ backend/src/main/java/com/back/ourlog/domain/comment/controller/CommentController.java
@@ -4,7 +4,9 @@ import com.back.ourlog.domain.comment.dto.CommentRequestDto;
 import com.back.ourlog.domain.comment.dto.CommentResponseDto;
 import com.back.ourlog.domain.comment.dto.CommentUpdateRequestDto;
 import com.back.ourlog.domain.comment.service.CommentService;
+import com.back.ourlog.domain.user.entity.User;
 import com.back.ourlog.global.common.dto.RsData;
+import com.back.ourlog.global.rq.Rq;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
@@ -21,11 +23,13 @@ import java.util.List;
 @Tag(name = "댓글 API")
 public class CommentController {
     private final CommentService commentService;
+    private final Rq rq;
     @PostMapping
     @Operation(summary = "댓글 등록")
     public ResponseEntity<RsData<CommentResponseDto>> writeComment(@RequestBody @Valid CommentRequestDto req) {
-        // 테스트 ver (우선 User가 있다고 가정)
-        CommentResponseDto res = commentService.write(req.getDiaryId(),null, req.getContent());
+        User user = rq.getCurrentUser();
+
+        CommentResponseDto res = commentService.write(req.getDiaryId(), user, req.getContent());
 
         return ResponseEntity.status(HttpStatus.CREATED)
                 .body(RsData.of("201-1", "댓글이 등록되었습니다.", res));
@@ -42,8 +46,12 @@ public class CommentController {
 
     @PutMapping()
     @Operation(summary = "댓글 수정")
-    public ResponseEntity<RsData<Void>> updateComment
-            (@RequestBody CommentUpdateRequestDto req) {
+    public ResponseEntity<RsData<Void>> updateComment(
+            @RequestBody CommentUpdateRequestDto req) {
+        User user = rq.getCurrentUser();
+
+        commentService.checkCanUpdate(user, req.getId());
+
         commentService.update(req.getId(), req.getContent());
 
         return ResponseEntity.status(HttpStatus.OK)
@@ -54,7 +62,12 @@ public class CommentController {
 
     @DeleteMapping("/{commentId}")
     @Operation(summary = "댓글 삭제")
-    public ResponseEntity<RsData<Void>> deleteComment(@PathVariable("commentId") int commentId) {
+    public ResponseEntity<RsData<Void>> deleteComment(
+            @PathVariable("commentId") int commentId) {
+        User user = rq.getCurrentUser();
+
+        commentService.checkCanDelete(user, commentId);
+
         commentService.delete(commentId);
 
         return ResponseEntity.status(HttpStatus.OK)
